requirements.txt
aiogram==3.5.0
aiohttp==3.8.4
python-dotenv==1.0.0
marzpy==0.0.6   # необязательно — я показал пример с прямыми запросами, но библиотека может помочь
